fixed matrices; fixed shape3D

+++ include/boglfw/renderOpenGL/Shape3D.h
@@ -80,7 +80,7 @@ private:
 	};
 	// line buffers
 	std::vector<s_vertex> buffer_;
-	std::vector<unsigned short> indices_;
+	std::vector<unsigned> indices_;
 	std::vector<unsigned> batches_;
 	glm::mat4 transform_ {1};
 	bool transformActive_ = false;
